# Improve unary broadcast efficiency for Diagonal type