This PR decreases the execution time of expressions like exp.(DM::Diagonal)
and abs.(DM::Diagonal). The improvement varies with the function and with
the size of DM, but it can be dramatic:

```julia
julia> M = Diagonal(rand(100));

julia> @btime exp.($DM);     # before
  41.856 μs (2 allocations: 78.20 KiB)

julia> @btime exp.($DM);     # after
  8.304 μs (2 allocations: 78.20 KiB)

julia> @btime abs.($DM);     # before
  179.532 ns (4 allocations: 960 bytes)

julia> @btime abs.($DM);     # after
  108.239 ns (2 allocations: 912 bytes)