@mbauman 2018-12-03
Thanks for the PR.  I don't think this is the way we want to go about this, though — this will break any gains from fusion.  While we do break fusion for some other types, we do it in favor of an O(1) optimization.  We can get these same gains _and allow fusion_ by implementing `copy(::Broadcasted{<:StructuredMatrixStyle})` or `copyto!(::AbstractArray, ::Broadcasted{<:StructuredMatrixStyle})`. Not sure which I would prefer at the moment.