@jlapeyre 2018-12-03
Ok. I noticed a performance loss in fusion in a similar experiment, but did not think to check for it here. But, you are correct, I suppose it will break every time:

Before the PR
```
julia> d1  = Diagonal(rand(10));

julia> @btime  sqrt.($d1);
  94.683 ns (4 allocations: 224 bytes)

julia> @btime 3 .* sqrt.($d1);
  139.394 ns (8 allocations: 320 bytes)
```
After the PR
```
julia> @btime  sqrt.($d1);                     #  faster
  79.028 ns (2 allocations: 176 bytes)

julia> @btime 3 .* sqrt.($d1);               # slower
  176.160 ns (9 allocations: 480 bytes)  
```