@mbauman 2018-12-03
Check out 

https://github.com/JuliaLang/julia/blob/7ea4542359b8629621a1f3cdb1f9aa97fdcfcde2/stdlib/LinearAlgebra/src/structuredbroadcast.jl#L104

It should be very simple to implement `copy(::Broadcasted{StructuredMatrixStyle{<:Diagonal}})` just like you have in this PR.  I think that should be simpler than `copyto!` due to those existing `copyto!` definitions (and the potential ambiguities with them).