@jlapeyre 2018-12-03
Thanks. I'll look into your suggestion... Now I remember. I tried this with `Eye` which *is* an `O(1)` optimization. But, I noticed it broke fusion, so I decided not to submit it to `FillArrays.jl`. I didn't know enough about broadcast to proceed. But, this is a good opportunity.