@mbauman 2018-12-03
In many cases, breaking fusion in favor of an O(1) algorithm is a big enough win that it's worth it.  That's what we do for ranges.  I'd think the same would be true for FillArray's `Eye`, but I've not tried it myself.  Note that I'd only do it in cases where I could _also_ return an `Eye`, targeting specific functions like `broadcasted(::typeof(*), ::Eye, ::Number)`, `broadcasted(::typeof(*), ::Number, ::Eye)`, etc.  But that's a different thing than here :)