@jlapeyre 2018-12-03
I can't quite recall the details. It was more dramatic that what I showed. But what I tried was indeed `broadcasted(::typeof(*), x::Number, ...)`.

Also, `broadcasted(::typeof(*), ::Eye, ::Number)` would in general return a `Diagonal(::Fill{T,1})`. Except you have to deal with `Inf * m::Eye`, etc.