@inkydragon 2025-01-15
We have moved the LinearAlgebra stdlib to an external repo: [JuliaLang/LinearAlgebra.jl
](https://github.com/JuliaLang/LinearAlgebra.jl)

If you think that this PR is still relevant, please open a new PR on the LinearAlgebra.jl repo.